renaming data server for gRPC switch
